GFF3_Annotation_extractor: move classes and functions into new 'annotation' module.
